feat: domain별로 validation들을 포함한 validator를 추가하도록 설계
